agent-backend/requirements.txt
fastapi
uvicorn
pydantic
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
